+++ include/gtirb_pprinter/ElfPrettyPrinter.hpp
@@ -116,8 +116,8 @@ protected:
   void printSectionFooterDirective(std::ostream& os,
                                    const gtirb::Section& addr) override;
 
-  // Print the `.size FunctionSymbol, . - FunctionSymbol` label that
-  // defines the size of the function symbol.
+  /** Print the `.size FunctionSymbol, . - FunctionSymbol` label that defines
+   * the size of the function symbol. */
   void printFunctionEnd(std::ostream& OS,
                         const gtirb::Symbol& FunctionSymbol) override;
 
@@ -149,8 +149,9 @@ protected:
   void printSymbolType(std::ostream& os, std::string& Name,
                        const aux_data::ElfSymbolInfo& SymbolInfo);
 
-  void printObjectSymbolSize(std::ostream& os, std::string& Name,
-                             const aux_data::ElfSymbolInfo& SymbolInfo);
+  /** Print .size directives for OBJECT and TLS symbols. */
+  void printDataSymbolSize(std::ostream& os, std::string& Name,
+                           const aux_data::ElfSymbolInfo& SymbolInfo);
 
   void printString(std::ostream& Stream, const gtirb::DataBlock& Block,
                    uint64_t Offset, bool NullTerminated = true) override;